move department name helper to concern
